Made latexmk_running_pids buffer local * As buffer local, latexmk bindings now work between buffers
This fixes a bug where latexmk mappings stop to work if you open a tex file and
starts latexmk (with <leader>ll), then open a new tex file in a different
buffer.  The reason it stopped working was simply that the list of running pids
was reset when a new buffer was opened.  If the list is defined as local to
a buffer, it works as it should.