First attempt at improved section movements
Adresses issue #76

I think this should work well (at least it does in my own test case).  However,
I was not able to split it up in the way it has been done for similar mappings.
That is, I tried to use <Plug> and to define the function in motion.vim, but
for some reason it never worked.

Also, the [] and ][ mappings could probably still be defined in a better way.