Fixed autocommand to kill latexmk * Ref. issue #87
The previous autocommand fails when the tex-buffers are deleted before one
exits vim.  This update changes the autocommand for kill_all_latexmk() to
a buffer specific autocommand for the event BufUnload.

Also add a simple check in kill_all_latexmk to prevent error messages if the
event is triggered after the buffer has been unloaded, which seems to happen
sometimes.